requirements.txt
emoji==2.8.0
extra-streamlit-components==0.1.56
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.1
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
regex==2023.8.8
requests==2.28.2
selenium==4.10.0
st-annotated-text==4.0.0
st-pages==0.4.4
streamlit==1.26.0
streamlit-aggrid==0.3.3
streamlit-agraph==0.0.45
streamlit-analytics==0.4.1
streamlit-authenticator==0.2.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-custom-notification-box==0.1.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
wordcloud==1.9.2

